RecordLogLoader: Add LoadRecordingAndExportAnim(), delete OutputLogLoader
