--- Editor/OutputLogLoaderEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43c29ac72344e2341986f84bde4625c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
