--- Runtime/CSharp/OutputLogLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0b5585f4f2bee04599ca34bad868599
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
